Move iteration portal item class
