+++ lib/helpers/views/cycleTimeDataProvider.js
@@ -1,7 +1,7 @@
 const {TreeItem, TreeItemCollapsibleState} = require('vscode')
-const moment = require('moment')
 const {model} = require('../../../model')
 const {commands} = require('../../commands')
+const IterationPortalItem = require('./iterationPortalItem')
 
 class CycleTimeTreeDataProvider {
   constructor(context) {
@@ -60,24 +60,6 @@ class CycleTimeTreeDataProvider {
 
   _getTreeIterations(year, month) {
     return this._portal[year][month].map(iteration => new TreeItem(iteration.label, TreeItemCollapsibleState.None))
-  }
-}
-
-class IterationPortalItem {
-  constructor(iteration) {
-    this.portalGroup = null
-    this.year = this._getYear(iteration.start)
-    this.month = this._getMonth(iteration.start)
-    this.iterationNumber = iteration.number
-    this.label = `${moment(iteration.start).format('DD-MMM-YYYY')} to ${moment(iteration.finish).format('DD-MMM-YYYY')}`
-  }
-
-  _getYear(startDate) {
-    return moment(startDate).format('YYYY')
-  }
-
-  _getMonth(startDate) {
-    return moment(startDate).format('MMMM')
   }
 }
